requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-common==1.1.28
azure-core==1.32.0
azure-functions==1.21.3
 azure-search-documents==11.5.2
azure-storage-blob==12.24.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cryptography==44.0.0
cymem==2.0.11
distro==1.9.0
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
langcodes==3.5.0
language_data==1.3.0
loguru==0.7.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.1
openai==1.59.8
packaging==24.2
preshed==3.0.9
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.2
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
text-unidecode==1.3
thinc==8.3.4
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
